requirements.txt
# Core Framework
streamlit>=1.28.0,<1.40.0

# AI & LLM
anthropic>=0.25.0

# PDF Processing
PyMuPDF>=1.23.0
pypdf>=3.0.0
Pillow>=10.0.0,<11.0.0

# Document Analysis & RAG
langchain>=0.1.0
langchain-community>=0.0.38
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# PDF Generation
reportlab>=4.0.0

# Text Processing
tiktoken>=0.5.0
python-dotenv>=1.0.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
tqdm>=4.65.0
plotly>=5.0.0

# Error Tracking & Logging
python-json-logger>=2.0.0

# Security
cryptography>=41.0.0

# Web Search (optional - for enhanced research)
tavily-python>=0.3.0

# Social Media Automation
tweepy>=4.14.0
apscheduler>=3.10.4
scipy>=1.11.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
